+++ exporter/opentelemetry-exporter-jaeger/CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
 ## Unreleased
-
 - Drop support for Python 3.4
   ([#1099](https://github.com/open-telemetry/opentelemetry-python/pull/1099))
+- Report instrumentation info
+  ([#1098](https://github.com/open-telemetry/opentelemetry-python/pull/1098))
 
 ## Version 0.12b0